m2-38 routing moved to app.tsx
